Created Utils to reduce copy-pasta
